pyproject.toml
[project]
name = "spd"
version = "0.0.1"
description = "Sparse Parameter Decomposition"
requires-python = "==3.12"
urls = { "Homepage" = "https://github.com/goodfire-ai/spd" }
license = { text = "MIT" }
readme = "README.md"
dependencies = [
    "torch<2.6.0",
    "torchvision",
    "pydantic",
    "wandb",
    "fire",
    "tqdm",
    "pytest",
    "ipykernel",
    "transformers",
    "jaxtyping",
    "einops",
    "matplotlib==3.9.1", # Avoid frequent pyright errors with new matplotlib versions
    "numpy",
    "python-dotenv",
    "wandb",
    "sympy",
    "streamlit",
    "streamlit-antd-components",
    "datasets",
    "pillow>=11.2.1",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pyright",
    "pre-commit",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["spd*"]

[tool.ruff]
line-length = 100
fix = true
ignore = [
    "F722", # Incompatible with jaxtyping
    "E731" # I think lambda functions are fine in several places
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.ruff.format]
# Enable reformatting of code snippets in docstrings.
docstring-code-format = true

[tool.ruff.isort]
known-third-party = ["wandb"]

[tool.pyright]
include = ["spd", "tests"]
exclude = ["**/wandb/**"]

strictListInference = true
strictDictionaryInference = true
strictSetInference = true
reportFunctionMemberAccess = true
reportUnknownParameterType = true
reportIncompatibleMethodOverride = true
reportIncompatibleVariableOverride = true
reportInconsistentConstructorType = true
reportOverlappingOverload = true
reportConstantRedefinition = true
reportImportCycles = true
reportPropertyTypeMismatch = true
reportMissingTypeArgument = true
reportUnnecessaryCast = true
reportUnnecessaryComparison = true
reportUnnecessaryContains = true
reportUnusedExpression = true
reportMatchNotExhaustive = true
reportShadowedImports = true
reportPrivateImportUsage = false
reportCallIssue = true

[tool.pytest.ini_options]
filterwarnings = [
    # https://github.com/google/python-fire/pull/447
    "ignore::DeprecationWarning:fire:59",
]
